Fine tuning the batch APIs and minor adjustments

+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/bulkimport/populator/AbstractWorkbookPopulator.java
@@ -46,7 +46,7 @@ public abstract class AbstractWorkbookPopulator implements WorkbookPopulator {
     }
 
     protected void writeLong(int colIndex, Row row, long value) {
-        row.createCell(colIndex).setCellValue(value);
+        row.createCell(colIndex).setCellValue((double) value);
     }
 
     protected void writeString(int colIndex, Row row, String value) {
